1. 选中标注后右击可以选择删除标注； 2. 修复拉伸标注后，不能及时更新的bug。

+++ OpenSeadragonAnnotation/src/Domain/AnnoMark.cs
@@ -23,16 +23,16 @@ namespace Domain
         public double height { get; set; }
         public string fill { get; set; }
         public string stroke { get; set; }
-        public int strokeWidth { get; set; }
+        public double strokeWidth { get; set; }
         //public string strokeDashArray { get; set; }
         public string strokeLineJoin { get; set; }
-        public int strokeMiterLimit { get; set; }
-        public int scaleX { get; set; }
-        public int scaleY { get; set; }
-        public int angle { get; set; }
+        public double strokeMiterLimit { get; set; }
+        public double scaleX { get; set; }
+        public double scaleY { get; set; }
+        public double angle { get; set; }
         public bool flipX { get; set; }
         public bool flipY { get; set; }
-        public int opacity { get; set; }
+        public double opacity { get; set; }
         //public int shadow { get; set; }
         public bool visible { get; set; }
         public string backgroundColor { get; set; }
